refactor(pack): refactor exchange individual
