+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/exchangeIndividual/add/add.routing.js
@@ -1,0 +1,11 @@
+import template from './add.html';
+import controller from './add.controller';
+
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('telecom.packs.pack.exchangeIndividual-add', {
+    url: '/exchangeIndividual/add',
+    template,
+    controller,
+    controllerAs: 'ctrl',
+  });
+};
