+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/exchangeIndividual/exchangeIndividual.html
