+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/exchangeIndividual/index.js
@@ -1,0 +1,32 @@
+import angular from 'angular';
+import ngOvhTelecomUniverseComponents from '@ovh-ux/ng-ovh-telecom-universe-components';
+import '@uirouter/angularjs';
+import '@ovh-ux/ng-translate-async-loader';
+import 'angular-translate';
+import 'ovh-api-services';
+
+import controller from './exchangeIndividual.controller';
+import template from './exchangeIndividual.html';
+
+const moduleName = 'PackExchangeIndividual';
+
+angular
+  .module(moduleName, [
+    ngOvhTelecomUniverseComponents,
+    'ngTranslateAsyncLoader',
+    'pascalprecht.translate',
+    'ovh-api-services',
+    'ui.router',
+  ])
+  .controller('PackExchangeIndividualCtrl', controller)
+  .run(/* @ngTranslationsInject:json ./translations */)
+  .run(
+    /* @ngInject */ ($templateCache) => {
+      $templateCache.put(
+        'app/telecom/pack/slots/exchangeIndividual/exchangeIndividual.html',
+        template,
+      );
+    },
+  );
+
+export default moduleName;
